Refactor: Restructure project for better organization and scalability
