--- src/config/loadEnv.js
@@ -1,3 +1,0 @@
-require("dotenv").config({
-  path: require("path").resolve(__dirname, "../../.env"),
-});
